move SupportsCatalogs and SupportsMetalakes in api module to core module
