+++ clients/client-java/src/main/java/com/datastrato/gravitino/client/api/SupportsCatalogs.java
@@ -16,10 +16,10 @@ import java.util.Map;
 
 /**
  * Client interface for supporting catalogs. It includes methods for listing, loading, creating,
- * altering and dropping catalogs. This class is forked from {@link
- * com.datastrato.gravitino.SupportsCatalogs}, which was designed for Gravitino server side. The *
- * difference is that this interface uses simple types instead of complex types in the method *
- * signatures, which will be easier for client side to use.
+ * altering and dropping catalogs. This class is forked from
+ * com.datastrato.gravitino.catalog.SupportsCatalogs, which was designed for Gravitino server side.
+ * The * difference is that this interface uses simple types instead of complex types in the method
+ * * signatures, which will be easier for client side to use.
  */
 @Evolving
 public interface SupportsCatalogs {
